% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_data_v2.3.R
\name{export_data_v2.3}
\alias{export_data_v2.3}
\title{Export of results}
\usage{
export_data_v2.3(
  occ_digital_voucher_file = "",
  occ_digital_voucher = NA,
  merge_unusable_data = FALSE,
  fields_to_merge = c("Ctrl_fieldNotes", "Ctrl_year", "Ctrl_stateProvince",
    "Ctrl_municipality", "Ctrl_locality", "Ctrl_countryCode", "Ctrl_eventDate",
    "Ctrl_habitat", "Ctrl_level0Name", "Ctrl_level1Name", "Ctrl_level2Name",
    "Ctrl_level3Name"),
  fields_to_compare = c("Ctrl_gbifID", "Ctrl_scientificName", "Ctrl_recordedBy",
    "Ctrl_recordNumber", "Ctrl_identifiedBy", "Ctrl_dateIdentified",
    "Ctrl_institutionCode", "Ctrl_collectionCode", "Ctrl_datasetName",
    "Ctrl_datasetName", "Ctrl_language", "wcvp_plant_name_id", "wcvp_taxon_rank",
    "wcvp_taxon_status", "wcvp_family", "wcvp_taxon_name", "wcvp_taxon_authors",
    "wcvp_reviewed", "wcvp_searchNotes"),
  fields_to_parse = c("Ctrl_gbifID", "Ctrl_bibliographicCitation", "Ctrl_language",
    "Ctrl_institutionCode", "Ctrl_collectionCode", "Ctrl_datasetName",
    "Ctrl_basisOfRecord", "Ctrl_catalogNumber", "Ctrl_recordNumber", "Ctrl_recordedBy",
    "Ctrl_occurrenceStatus", "Ctrl_eventDate", "Ctrl_year", "Ctrl_month", "Ctrl_day",
    "Ctrl_habitat", "Ctrl_fieldNotes", "Ctrl_eventRemarks", "Ctrl_countryCode",
    "Ctrl_stateProvince", "Ctrl_municipality", "Ctrl_county", "Ctrl_locality",
    "Ctrl_level0Name", "Ctrl_level1Name", "Ctrl_level2Name", 
     "Ctrl_level3Name",
    "Ctrl_identifiedBy", "Ctrl_dateIdentified", "Ctrl_scientificName",
    "Ctrl_decimalLatitude", "Ctrl_decimalLongitude", "Ctrl_nameRecordedBy_Standard",
    "Ctrl_recordNumber_Standard", "Ctrl_key_family_recordedBy_recordNumber",
    "Ctrl_geospatial_quality", "Ctrl_verbatim_quality", "Ctrl_moreInformativeRecord",
    "Ctrl_coordinates_validated_by_gbif_issue", "wcvp_plant_name_id", "wcvp_taxon_rank",
    "wcvp_taxon_status", "wcvp_family", "wcvp_taxon_name", "wcvp_taxon_authors",
    "wcvp_reviewed", 
     "wcvp_searchNotes", "parseGBIF_digital_voucher",
    "parseGBIF_duplicates", "parseGBIF_num_duplicates",
    "parseGBIF_non_groupable_duplicates", "parseGBIF_duplicates_grouping_status",
    "parseGBIF_unidentified_sample", "parseGBIF_sample_taxon_name",
    "parseGBIF_sample_taxon_name_status", "parseGBIF_number_taxon_names",
    "parseGBIF_useful_for_spatial_analysis", "parseGBIF_decimalLatitude",
    "parseGBIF_decimalLongitude", "parseGBIF_dataset_result"),
  silence = TRUE
)
}
\arguments{
\item{occ_digital_voucher_file}{CSV fila result of function select_digital_voucher()$occ_digital_voucher}

\item{occ_digital_voucher}{data frame result of function select_digital_voucher()$occ_digital_voucher}

\item{merge_unusable_data}{include records unique collection events incomplete in merge processing}

\item{fields_to_merge}{fields to merge}

\item{fields_to_compare}{fields to compare content frequency}

\item{fields_to_parse}{all fields}

\item{silence}{if TRUE does not display progress messages}
}
\value{
list with 10 data frames
\itemize{
\item \strong{all_data} All records processed, merged Unique collection events complete / incomplete and their duplicates
\item \strong{useable_data_merge} Merged Unique collection events complete
\item \strong{useable_data_raw} Raw Unique collection events complete
\item \strong{duplicates} Duplicates of unique collection events complete / incomplete
\item \strong{unusable_data_merge} Merged Unique collection events incomplete,
It is NA if merge_unusable_data is FALSE.
\item \strong{unusable_data_raw} Raw Unique collection events incomplete
\item \strong{parseGBIF_general_summary}
\item \strong{parseGBIF_merge_fields_summary}
\item \strong{parseGBIF_merge_fields_summary_useable_data}
\item \strong{parseGBIF_merge_fields_summary_unusable_data} It is NA if merge_unusable_data is FALSE
}
}
\description{
For each unique collection event key, complete or incomplete,
outputs will be created which combine information from duplicate records and generate a
single unique collection event record to replace them.
The main output fields relating to taxonomic identification and geographic coordinates:
\itemize{
\item parseGBIF_sample_taxon_name = scientific name chosen as taxonomic identification for unique collection event
\item parseGBIF_number_taxon_names = number of scientific names found in duplicates of unique collection event
\item parseGBIF_sample_taxon_name_status = status of choice of 'identified', 'divergent identifications', 'unidentified'
\item parseGBIF_unidentified_sample = if unique collection event has taxonomic identification
\item parseGBIF_decimalLatitude = latitude in decimal degrees
\item parseGBIF_decimalLongitude = longitude in decimal degrees
\item parseGBIF_useful_for_spatial_analysis = whether the coordinates are useful for spatial analysis.
\strong{How is the taxon binomial attributed to the unique collection event selected?}
}
\enumerate{
\item Where the unique collection event key is complete:
The accepted TAXON_NAME selected is that which is most frequently applied to the duplicate vouchers at or below the rank of species.
Where two named are applied with equal frequency then a mechanical approach, using alphabetical order, is applied, the first listed TAXON_NAME being chosen.
Where there is no identification, at or below the rank of species, then the unique collection event, the unique collection event is indicated as unidentified.
\item Where the unique collection event key is incomplete:
Where the unique collection event key is incomplete, then each record is treated as a unique collection event. If there is no identification, at or below the rank of species, then the unique collection event is classified as unidentified.
__Geospatial information __
If the master voucher does not have geographic coordinates, we will seek coordinates from the duplicate records associated with it.
Finally, the records are separated into three sets of data:
\strong{useable_data} - Where unique collection event with taxonomic identification and geographic coordinates are complete. This represents the useable dataset.
\strong{unusable_data}  - Where unique collection event without taxonomic identification and/or geographic coordinates.
\strong{duplicates} The duplicates of unique collection events complete / incomplete
}

With this, it is possible to perform:
Merge information between fields of duplicates of a unique collection event to create a synthetic record for each unique collection event,
Compare the frequency of content in fields
Generate a work package summary

For each complete unique collection event key, data fields that are empty in the digital voucher record will be populated with data from the respective duplicates.
During content merging, we indicate fields associated with the description, location, and data of the unique collection event.
By default, fields_to_merge parameter of export_data function contains:
\itemize{
\item Ctrl_fieldNotes
\item Ctrl_year
\item Ctrl_stateProvince
\item Ctrl_municipality
\item Ctrl_locality
\item Ctrl_countryCode
\item Ctrl_eventDate
\item Ctrl_habitat
\item Ctrl_level0Name
\item Ctrl_level1Name
\item Ctrl_level2Name
\item Ctrl_level3Name
}
}
\details{
Each data frame should be used as needed
}
\examples{
\donttest{
help(export_data)

results <- export_data(occ_digital_voucher_file = file.occ_digital_voucher,
                       merge_unusable_data = TRUE)

names(results)

results$parseGBIF_general_summary
results$parseGBIF_merge_fields_summary
results$parseGBIF_merge_fields_summary_complete
NROW(results$all_data)
NROW(results$unique_collection_event_complete_merge)
NROW(results$unique_collection_event_incomplete_raw)
NROW(results$duplicates)

}
}
\seealso{
\code{\link[ParsGBIF]{batch_checkName_wcvp}}, \code{\link[ParsGBIF]{extract_gbif_issue}}
}
\author{
Pablo Hendrigo Alves de Melo,
Nadia Bystriakova &
Alexandre Monro
}
