% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_coordinates.R
\name{parse_coordinates}
\alias{parse_coordinates}
\title{Checking coordinates, centroids, artificial points and others}
\usage{
parse_coordinates(
  occ = NA,
  file_occ = NA,
  iso2_field_name = "Ctrl_countryCode",
  path_centroids = "https://raw.githubusercontent.com/pablopains/parseGBIF/main/dataRaw",
  scale = 110
)
}
\arguments{
\item{occ}{GBIF occurrence table with selected columns as select_gbif_fields(columns = 'standard')}

\item{iso2_field_name}{indicates the name of the field with ISO2 code of the countries}
}
\value{
list with two data frames, occ, with the original data set plus two columns, parseGBIF_countryCode_ISO3 and
parseGBIF_countryName_en and countrycodelist, with the list of countries found with all the columns of countrycode::codelist
}
\description{
Checking coordinates, centroids, artificial points, also makes use of basic functions of CoordinateCleaner and other packages
}
\details{
Returns the last name
}
\examples{
\donttest{
help(standardize_country_from_iso2)

occ <- prepare_gbif_occurrence_data(gbif_occurrece_file =  'https://raw.githubusercontent.com/pablopains/parseGBIF/main/dataGBIF/Achatocarpaceae/occurrence.txt',
                                    columns = 'standard')
x <- standardize_country_from_iso2(occ = occ,
                                     iso2_field_name = 'Ctrl_countryCode',
                                     return_fields = c('iso3c','country.name.en'))
colnames(x$occ)
head(x$countrycodelist)
}
}
\seealso{
\code{\link[parseGBIF]{prepare_gbif_occurrence_data}}, \code{\link[parseGBIF]{download_gbif_data_from_doi}}
}
\author{
Pablo Hendrigo Alves de Melo,
Nadia Bystriakova &
Alexandre Monro
}
