% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseGBIF_app.R
\name{parseGBIF_app}
\alias{parseGBIF_app}
\title{parseGBIF App}
\usage{
parseGBIF_app()
}
\value{
CSV files
}
\description{
parseGBIF App
}
\examples{
\donttest{
parseGBIF::parseGBIF_app()
}
}
\seealso{
\code{\link[ParsGBIF]{batch_checkName_wcvp}}, \code{\link[ParsGBIF]{extract_gbif_issue}}
}
\author{
Pablo Hendrigo Alves de Melo,
}
