% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_diagnosis_collection_event_key_quality.R
\name{pre_diagnosis_collection_event_key_quality}
\alias{pre_diagnosis_collection_event_key_quality}
\title{Pre-diagnosis of collection event key quality}
\usage{
pre_diagnosis_collection_event_key_quality(
  occ = NA,
  n_occ_min = 10,
  n_show = 50
)
}
\arguments{
\item{occ}{GBIF occurrence table with selected columns as select_gbif_fields(columns = 'standard')}

\item{n_occ_min}{Minimum records per collector per collection}

\item{n_show}{Inconsistency number to list}
}
\value{
Text report
}
\description{
Pre-diagnosis of collection event key quality
}
\details{
Pre-diagnosis of collection event key quality
}
\examples{
\donttest{

library(ParsGBIF)

help(download_gbif_data_from_doi)

download_gbif_data_from_doi(gbif_doi_url='https://www.gbif.org/occurrence/download/0151470-230224095556074',
                            folder = 'c://dataGBIF//Achatocarpaceae')

}
}
\seealso{
\code{\link[parseGBIF]{prepare_gbif_occurrence_data}}, \code{\link[parseGBIF]{extract_gbif_issue}}
}
\author{
Pablo Hendrigo Alves de Melo,
Nadia Bystriakova &
Alexandre Monro
}
