% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_gbif_occurrence_data.R
\name{prepare_gbif_occurrence_data}
\alias{prepare_gbif_occurrence_data}
\title{Preparing occurrence data downloaded from GBIF for use by parseGBIF}
\usage{
prepare_gbif_occurrence_data(gbif_occurrece_file = "", columns = "standard")
}
\arguments{
\item{gbif_occurrece_file}{The name of the file with occurrence data downloaded from GBIF}

\item{columns}{Character vector of strings to indicate column names of the GBIF occurrence file.
Use 'standard' to select basic columns for use in the package, 'all' to select all available columns.
The default is 'standard'}
}
\value{
data.table with renamed selected fields with prefix "Ctrl_"
}
\description{
Prepare occurrence data downloaded from GBIF to be used by parseGBIF functions
}
\details{
Select data fields and rename field names prefixed with "Ctrl_"
}
