% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_digital_voucher.R
\name{select_digital_voucher}
\alias{select_digital_voucher}
\title{Selecting the master digital voucher (FAST)}
\usage{
select_digital_voucher(
  occ = NA,
  occ_gbif_issue = NA,
  occ_wcvp_check_name = NA,
  occ_collectorsDictionary = NA,
  enumOccurrenceIssue = NA,
  silence = TRUE
)
}
\arguments{
\item{occ}{GBIF occurrence table with selected columns as select_gbif_fields(columns = 'standard')}

\item{occ_gbif_issue}{result of function extract_gbif_issue()$occ_gbif_issue}

\item{occ_wcvp_check_name}{result of function batch_checkName_wcvp()$occ_wcvp_check_name}

\item{occ_collectorsDictionary}{result of function update_collectorsDictionary()$occ_collectorsDictionary}

\item{enumOccurrenceIssue}{An enumeration of validation rules for single occurrence records by GBIF file, if NA, will be used, data(EnumOccurrenceIssue)}

\item{silence}{if TRUE does not display progress messages}
}
\value{
list with two data frames:
\itemize{
\item occ_digital_voucher: full output dataset (occ_all) with parseGBIF_* fields, plus WCVP sample columns
\item occ_results: only processing/result fields
}
}
\description{
Fast rewrite of the original function. Groups duplicates by Ctrl_key_family_recordedBy_recordNumber,
computes record completeness + geospatial quality, selects a master digital voucher per group,
sets sample identification fields, and chooses coordinates for spatial analysis.
}
