% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/select_digital_voucher_and_sample_identification.R
\name{select_digital_voucher_and_sample_identification}
\alias{select_digital_voucher_and_sample_identification}
\title{Select a sample among available duplicates}
\usage{
select_digital_voucher_and_sample_identification(
  occ = NA,
  occ_gbif_issue = NA,
  occ_wcvp_check_name = NA,
  occ_collectorsDictionary = NA,
  enumOccurrenceIssue = NA
)
}
\arguments{
\item{occ}{GBIF occurrence table with selected columns as select_gbif_fields(columns = 'standard')}

\item{occ_gbif_issue}{= result of function extract_gbif_issue()$occ_gbif_issue}

\item{occ_wcvp_check_name}{= result of function batch_checkName_wcvp()$occ_wcvp_check_name}

\item{occ_collectorsDictionary}{= result of function update_collectorsDictionary()$occ_collectorsDictionary}

\item{enumOccurrenceIssue}{An enumeration of validation rules for single occurrence records by GBIF file, if NA, will be used, data(EnumOccurrenceIssue)}
}
\value{
list with two data frames: occ_digital voucher_and:
occ_digital_voucher_and_sample_identification, only with selection result fields and
occ_join_results, with all data processing fields.
}
\description{
To group duplicates and choose the digital voucher:
\enumerate{
\item If the key for grouping duplicates is complete with collector information and collection number, sample duplicates can be grouped. In this case, the voucher with the highest score is selected among the duplicates in the sample.
\item If the key to group duplicates is incomplete, sample duplicates cannot be grouped due to missing collector information and/or collection number. In this case, each record is considered a sample, without duplicates, and a voucher is selected for each sample.
}

How is the information score calculated?

moreInformativeRecord = sum of textual quality + quality of geospatial information.

How is the quality of textual information calculated?

The Text quality is the sum of the number of flags with text quality equal to TRUE.

Is there information about the collector?
Is there information about the collection number?
Is there information about the year of collection?
Is there information about the institution code?
Is there information about the catalog number?
Is there information about the collection site?
Is there information about the municipality of collection?
Is there information about the state/province of collection?
Is there information about the bibliographic citation?

How is the quality of geospatial information calculated?

The quality of geospatial information is based on geographic issues made available by GBIF.

GIBF issues on the quality of geospatial information were classified into three levels.
\itemize{
\item Not applicable, with selection_score equal to 0
\item Does not affect coordinating accuracy, with selection_score equal to -1
\item Potentially affect coordinate accuracy, with selection_score equal to -3
\item Records to be excluded from spatial analysis, with selection_score equal to -9
}

How is the taxonomic identification of the sample chosen?
\enumerate{
\item When the key to group the duplicates is complete:
The accepted TAXON_NAME identified at or below the specified level and the most frequent
among the duplicates is chosen.
}

In case of a tie in frequency, in alphabetical order, the first accepted TAXON_NAME
identified up to or below the specific level is chosen.

If there is no identification, equal to or less than the specific level, for the sample,
the sample is indicated as unidentified.
\enumerate{
\item When the key to group the duplicates is incomplete:
If so, the accepted TAXON_NAME identified at or below the specified level is used.
If there is no identification, equal to or less than the specific level,
the sample is indicated as unidentified.
}
}
\details{
\itemize{
\item matchStatusDuplicates - "matched", "unmatched: no recordedBy and no recordNumber",
"unmatched: no recordNumber" or "unmatched: no recordedBy"
\item numberTaxonNamesSample -  count of the different accepted scientific names,
identified up to or below the specific level, listed in the sample duplicates, or Zero,
if there is no identification, equal to or below the specific level, for the sample.
\item sampleTaxonName - TAXON_name accepted and identified up to or below the specific level
selected for the sample.
\item sampleIdentificationStatus - 'Identified', 'divergent identifications', or 'unidentified'
}
}
\examples{
\donttest{
help(select_digital_voucher_and_sample_identification)

head(occ)
head(res_gbif_issue$occ_gbif_issue)
head(res_checkName_wcvp$occ_wcvp_check_name)
head(res_collectorsDictionary$occ_collectorsDictionary)
res_digital_voucher_and_sample_identification <- select_digital_voucher_and_sample_identification(occ = occ,
                                                                                                  occ_gbif_issue = res_gbif_issue$occ_gbif_issue,
                                                                                                  occ_wcvp_check_name = res_checkName_wcvp$occ_wcvp_check_name,
                                                                                                  occ_collectorsDictionary = res_collectorsDictionary$occ_collectorsDictionary,
                                                                                                  enumOccurrenceIssue = EnumOccurrenceIssue)

names(res_digital_voucher_and_sample_identification)

head(res_digital_voucher_and_sample_identification$occ_digital_voucher_and_sample_identification)
colnames(res_digital_voucher_and_sample_identification$occ_digital_voucher_and_sample_identification)

head(res_digital_voucher_and_sample_identification$occ_join_results)
colnames(res_digital_voucher_and_sample_identification$occ_join_results)
}
}
\seealso{
\code{\link[ParsGBIF]{batch_checkName_wcvp}}, \code{\link[ParsGBIF]{extract_gbif_issue}}
}
\author{
Pablo Hendrigo Alves de Melo,
Nadia Bystriakova &
Alexandre Monro
}
