% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wcvp_check_name_batch.R
\name{wcvp_check_name_batch}
\alias{wcvp_check_name_batch}
\title{In batch, use the WCVP database to check accepted names and update synonyms}
\usage{
wcvp_check_name_batch(
  occ = NA,
  wcvp_names = "",
  if_author_fails_try_without_combinations = TRUE,
  wcvp_selected_fields = "standard",
  silence = TRUE
)
}
\arguments{
\item{occ}{GBIF occurrence table with selected columns as select_gbif_fields(columns = 'standard')}

\item{wcvp_names}{get data frame in parseGBIF::wcvp_get_data(read_only_to_memory = TRUE)$wcvp_names}

\item{if_author_fails_try_without_combinations}{option for partial verification of the authorship of the species.}

\item{wcvp_selected_fields}{WCVP fields selected as return, 'standard' basic columns, 'all' all available columns.}

\item{silence}{if TRUE does not display progress messages}
}
\value{
list with two data frames:
\itemize{
\item \code{summary}: summary of name checking results
\item \code{occ_wcvp_check_name}: occurrence data with WCVP fields
}
}
\description{
Species' names can be checked against WCVP database one by one, or in a batch mode.
To verify individual names, the function wcvp_check_name is used.
}
