% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wcvp_get_data.R
\name{wcvp_get_data}
\alias{wcvp_get_data}
\title{Get local WCVP database}
\usage{
wcvp_get_data(path_data, load_distribution = FALSE, silence = FALSE)
}
\arguments{
\item{path_data}{Character string. The path to the local directory containing the unzipped WCVP CSV files, or the path to the \code{wcvp.zip} file.}

\item{load_distribution}{Logical. TRUE to also load the file with geographical distribution of species.}

\item{silence}{Logical. TRUE to suppress progress messages.}
}
\value{
list with two data frames:
\itemize{
\item \code{wcvp_names}: taxonomic names database
\item \code{wcvp_distribution}: geographical distribution data (if load_distribution = TRUE)
}
}
\description{
Load the World Checklist of Vascular Plants (WCVP) database from a local directory or zip file.
}
\details{
To maximize performance and reliability, this function is designed to read the WCVP dataset from a local download rather than pulling directly from the KEW SFTP server during execution. You can download the latest version from: http://sftp.kew.org/pub/data-repositories/WCVP/

This space contains data resources publicly accessible to the user 'anonymous'. No password required for access. Use of data made available via this site may be subject to legal and licensing restrictions. The README in the top-level directory for each data resource provides specific information about its terms of use.
}
\examples{
\donttest{
# load package
library(parseGBIF)

help(wcvp_get_data)

# Point to the directory where you extracted the WCVP download
path_data <- "C:/parseGBIF/dataWCVP"

wcvp <- wcvp_get_data(path_data = path_data,
                      load_distribution = TRUE)

names(wcvp)

head(wcvp$wcvp_names)
colnames(wcvp$wcvp_names)

head(wcvp$wcvp_distribution)
colnames(wcvp$wcvp_distribution)
}

}
\seealso{
\code{\link[parseGBIF]{wcvp_check_name}}, \code{\link[parseGBIF]{wcvp_check_name_batch}}
}
\author{
Pablo Hendrigo Alves de Melo,
Nadia Bystriakova &
Alexandre Monro
}
