% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wcvp_get_dat_v2.R
\name{wcvp_get_data_v2}
\alias{wcvp_get_data_v2}
\title{Get WCVP database}
\usage{
wcvp_get_data_v2(
  url_source = "http://sftp.kew.org/pub/data-repositories/WCVP/",
  read_only_to_memory = FALSE,
  path_results = "C:/parseGBIF",
  update = FALSE,
  load_distribution = FALSE,
  load_rda_data = FALSE
)
}
\arguments{
\item{url_source}{http://sftp.kew.org/pub/data-repositories/WCVP/}

\item{read_only_to_memory}{TRUE to in-memory read-only, not writing a copy to local disk}

\item{path_results}{download destination folder, if read_only_to_memory FALSE}

\item{update}{TRUE to update and load files, FALSE to keep local version and load files, if read_only_to_memory FALSE}

\item{load_distribution}{TRUE to load file with geographical distribution of species, if read_only_to_memory FALSE}

\item{load_rda_data}{Load the WCVP name database from the rda file distributed with the package. To ensure updates, it is recommended to reinstall the package frequently.}
}
\value{
list with two data frames: wcvp_names and wcvp_distribution
}
\description{
Download World Checklist of Vascular Plants (WCVP) database
}
\details{
http://sftp.kew.org/pub/data-repositories/WCVP/ This is the public SFTP (Secure File Transfer Protocol) site of the Royal Botanic Gardens, Kew. This space contains data resources publicly accessible to the user `anonymous'.  No password required for access. Use of data made available via this site may be subject to legal and licensing restrictions. The README in the top-level directory for each data resource provides specific information about its terms of use.
}
\examples{
\donttest{
# load package
library(parseGBIF)

help(wcvp_get_data)

# Download wcvp database to local disk
path_data <- tempdir() # you can change this folder

wcvp <- wcvp_get_data(url_source = 'http://sftp.kew.org/pub/data-repositories/WCVP/',
                 read_only_to_memory = FALSE,
                 path_results = path_data,
                 update = FALSE,
                 load_distribution = TRUE)

names(wcvp)

head(wcvp$wcvp_names)
colnames(wcvp$wcvp_names)

head(wcvp$wcvp_distribution)
colnames(wcvp$wcvp_distribution)
}

}
\seealso{
\code{\link[parseGBIF]{wcvp_check_name}}, \code{\link[parseGBIF]{wcvp_check_name_batch}}
}
\author{
Pablo Hendrigo Alves de Melo,
Nadia Bystriakova &
Alexandre Monro
}
