% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wcvp_prepare_index.R
\name{wcvp_prepare_index}
\alias{wcvp_prepare_index}
\title{Prepare an index for WCVP name lookup (speed optimization)}
\usage{
wcvp_prepare_index(wcvp_names, overwrite = FALSE)
}
\arguments{
\item{wcvp_names}{data.frame as loaded from WCVP (must contain TAXON_NAME_U, TAXON_AUTHORS_U)}

\item{overwrite}{if TRUE, rebuild index even if already present}
}
\value{
wcvp_names with lookup index attached (attributes) and helper key columns added
}
\description{
Prepare an index for WCVP name lookup (speed optimization)
}
