% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collectors_get_name.R
\name{collectors_get_name}
\alias{collectors_get_name}
\title{Get the last name of the main collector in recordedBy field}
\usage{
collectors_get_name(x)
}
\arguments{
\item{x}{recordedBy field}
}
\value{
last name of the main collector
}
\description{
Returns the last name of the main collector
}
\details{
Returns the last name of the main collector in recordedBy field
}
\examples{
\donttest{
help(collectors_get_name)

collectors_get_name('Melo, P.H.A & Monro, A.')

collectors_get_name('Monro, A. & Melo, P.H.A')
}

}
\seealso{
\code{\link[parseGBIF]{collectors_prepare_dictionary}}, \code{\link[parseGBIF]{collectors_update_dictionary}}
}
\author{
Pablo Hendrigo Alves de Melo,
Nadia Bystriakova &
Alexandre Monro
}
