% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_gbif_data_from_doi.R
\name{download_gbif_data_from_doi}
\alias{download_gbif_data_from_doi}
\title{Download GBIF occurrence data from DOI}
\usage{
download_gbif_data_from_doi(
  gbif_doi_url,
  folder = "",
  keep_only_occurrence_file = TRUE,
  overwrite = TRUE
)
}
\arguments{
\item{gbif_doi_url}{The url of the GBIF DOI}

\item{folder}{Save folder}

\item{keep_only_occurrence_file}{Keep only occurrence.txt file}

\item{overwrite}{overwrite files}

\item{subfolder}{Save in subfolder}
}
\value{
list of downloaded and unzipped files
}
\description{
Download and unzip GBIF occurrence data from DOI to be used by ParsGBIF functions
}
\details{
Download GBIF occurrence data downloaded from DOI
}
\examples{
\donttest{

library(ParsGBIF)

help(download_gbif_data_from_doi)

download_gbif_data_from_doi(gbif_doi_url='https://www.gbif.org/occurrence/download/0151470-230224095556074',
                            folder = 'c://dataGBIF//Achatocarpaceae')

}
}
\seealso{
\code{\link[parseGBIF]{prepare_gbif_occurrence_data}}, \code{\link[parseGBIF]{extract_gbif_issue}}
}
\author{
Pablo Hendrigo Alves de Melo,
Nadia Bystriakova &
Alexandre Monro
}
