% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_gbif_issue.R
\name{extract_gbif_issue}
\alias{extract_gbif_issue}
\title{Extracting GBIF issues}
\usage{
extract_gbif_issue(occ = NA, enumOccurrenceIssue = NA)
}
\arguments{
\item{occ}{GBIF occurrence table with selected columns as select_gbif_fields(columns = 'standard')}

\item{enumOccurrenceIssue}{An enumeration of validation rules for single occurrence records by GBIF file, if NA, will be used, data(EnumOccurrenceIssue)}
}
\value{
list with two data frames: summary, with the frequency of issues in the records
and occ_gbif_issue, with issues in columns with TRUE or FALSE for each record.
}
\description{
Extract GBIF validation rules for occurrence records

GBIF recognises and documents several issues relating to the data fields for an individual record.
The issue field stores terms that represent an enumeration of GBIF validation rules.
Issues can lead to errors or unexpected data. The issues fields are therefore a valuable source of information
when assessing the quality of a record. In order to help GBIF and the data publishers improve the data,
GBIF flag records with various issues that they have encountered. These issues can be  used as filters applied
to occurrence searches. Not all issues indicate bad data, some flagthe fact that GBIF has altered values during
processing. The values of EnumOccurrenceIssue will be used by the function extract_gbif_issue as a model to tabulate
the GBIF issues of each record, individualizing them, in columns.TRUE or FALSE, flagging whether the issue applies or
not for each record.
}
\details{
https://gbif.github.io/parsers/apidocs/org/gbif/api/vocabulary/OccurrenceIssue.html
}
\examples{
\donttest{

library(ParsGBIF)

help(extract_gbif_issue)

occ_file <- 'https://raw.githubusercontent.com/pablopains/ParsGBIF/main/dataGBIF/Achatocarpaceae/occurrence.txt'

occ <- prepare_gbif_occurrence_data(gbif_occurrece_file = occ_file,
                                    columns = 'standard')

data(EnumOccurrenceIssue)

colnames(EnumOccurrenceIssue)

head(EnumOccurrenceIssue)

occ_gbif_issue <- extract_gbif_issue(occ = occ)

names(occ_gbif_issue)

head(occ_gbif_issue$summary)

colnames(occ_gbif_issue$occ_gbif_issue)

head(occ_gbif_issue$occ_gbif_issue)
}

}
\seealso{
\code{\link[ParsGBIF]{prepare_gbif_occurrence_data}}, \code{\link[ParsGBIF]{select_gbif_fields}}
}
\author{
Pablo Hendrigo Alves de Melo,
Nadia Bystriakova &
Alexandre Monro
}
