% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flora_funga_brasil_check_name.R,
%   R/flora_funga_brasil_get_data.R
\name{flora_funga_brasil_get_data}
\alias{flora_funga_brasil_get_data}
\alias{flora_funga_brasil_check_name}
\title{Check species names against Funga  do Brazil database}
\usage{
flora_funga_brasil_check_name(
  searchedName = "Gardnerina angustata",
  flora_funga_brasil = "",
  if_author_fails_try_without_combinations = TRUE,
  silence = TRUE
)

flora_funga_brasil_get_data(
  url_source = "http://ipt.jbrj.gov.br/jbrj/archive.do?r=lista_especies_flora_brasil",
  path_results = ""
)
}
\arguments{
\item{searchedName}{= searched name}

\item{flora_funga_brasil}{Flora e Funga  do Brazil database}

\item{url_source}{= "http://ipt.jbrj.gov.br/jbrj/archive.do?r=lista_especies_flora_brasil"}

\item{path_results}{download destination folder}
}
\value{
Flora e Funga  do Brazil database

list with two data frames: wcvp_names and wcvp_distribution
}
\description{
Check species names against Flora e Funga do Brazil database

Download Flora e Funga  do Brazil database
}
\details{
....

http://sftp.kew.org/pub/data-repositories/WCVP/ This is the public SFTP (Secure File Transfer Protocol) site of the Royal Botanic Gardens, Kew. This space contains data resources publicly accessible to the user `anonymous'.  No password required for access. Use of data made available via this site may be subject to legal and licensing restrictions. The README in the top-level directory for each data resource provides specific information about its terms of use.
}
\examples{
\donttest{
# load package
library(parseGBIF)

help(flora_funga_brasil_get_data)

# Download database to local disk
path_data <- tempdir() # you can change this folder

flora_funga_brasil <- flora_funga_brasil_get_data(path_results = path_data,
                 update = FALSE,
                 load_distribution = TRUE)
}

\donttest{
# load package
library(parseGBIF)

help(flora_funga_brasil_get_data)

# Download database to local disk
path_data <- tempdir() # you can change this folder

flora_funga_brasil <- flora_funga_brasil_get_data(path_results = path_data,
                 update = FALSE,
                 load_distribution = TRUE)
}

}
\seealso{
\code{\link[parseGBIF]{wcvp_check_name}}, \code{\link[parseGBIF]{wcvp_check_name_batch}}

\code{\link[parseGBIF]{wcvp_check_name}}, \code{\link[parseGBIF]{wcvp_check_name_batch}}
}
\author{
Pablo Hendrigo Alves de Melo,
Nadia Bystriakova &
Alexandre Monro
}
