% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_collection_event_key.R
\encoding{UTF-8}
\name{generate_collection_event_key}
\alias{generate_collection_event_key}
\title{Generating the collection event key}
\usage{
generate_collection_event_key(
  occ = NA,
  collectorDictionary_checked_file = NA,
  collectorDictionary_checked = NA,
  collectorDictionary_file =
    "https://raw.githubusercontent.com/pablopains/parseGBIF/main/collectorDictionary/CollectorsDictionary.csv",
  collectorDictionary = NA,
  silence = TRUE
)
}
\arguments{
\item{occ}{GBIF occurrence table with selected columns as select_gbif_fields(columns = 'standard')}

\item{collectorDictionary_checked_file}{Verified collector dictionary file - point to a file on your local disk (use file or data frame)}

\item{collectorDictionary_checked}{Verified collector dictionary data frame (use file or data frame)}

\item{collectorDictionary_file}{Collector dictionary file - point to a file on your local disk or upload via git at https://raw.githubusercontent.com/pablopains/parseGBIF/main/collectorDictionary/CollectorsDictionary.csv.}

\item{silence}{if TRUE does not display progress messages}
}
\value{
list with three data frames:
occ_collectorsDictionary, with update result fields only,
summary and
CollectorsDictionary_add, with new collectors that can be added to the
collector dictionary that can be reused in the future.
}
\description{
This generates a key to identify the physical and digital duplicates, of a given collection event.
It combines the primary collector's surname, the collector's number and the botanical family, a key is created
(family + recordByStandardized + recordNumber_Standard) that allows grouping the duplicates of the same unique
collection event.

It also identifiesnew collectors to be added to the collector dictionary and that can be reused in the future.

Include recordedByStandardized field with verified main collector's last name.
Include recordNumber_Standard field with only numbers from recordNumber.
Create the collection event key to group duplicates in the key_family_recordedBy_recordNumber field,
composed of the fields: family + recordedByStandardized + recordNumber_Standard.
}
\details{
Fields created for each incident record:
nameRecordedBy_Standard,
recordNumber_Standard,
key_family_recordedBy_recordNumber,
key_year_recordedBy_recordNumber
}
\examples{
\donttest{
collectorsDictionaryFromDataset <- prepare_lastNameRecordedBy(occ=occ,
                                                              collectorDictionary_checked_file='collectorDictionary_checked.csv')

names(collectorsDictionaryFromDataset)

}

}
\seealso{
\code{\link[parseGBIF]{collectors_get_name}}, \code{\link[parseGBIF]{prepare_collectorsDictionary}}
}
\author{
Pablo Hendrigo Alves de Melo,
Nadia Bystriakova &
Alexandre Monro
}
