% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_centroids.R
\name{get_centroids}
\alias{get_centroids}
\title{Load or generate table with centroids}
\usage{
get_centroids(
  path_centroids = "https://raw.githubusercontent.com/pablopains/parseGBIF/main/dataRaw"
)
}
\arguments{
\item{path_centroids}{path to the centroids file, default 'https://raw.githubusercontent.com/pablopains/parseGBIF/main/dataRaw'
if not provided (NA), the table will be generated and loaded only into memory.}
}
\value{
Table with centroids for levels 0, 1 and 2
}
\description{
load or generate table with centroids for levels 0, 1 and 2,
countries, states and municipalities, in the world
}
\details{
Returns the table with centroids
}
\examples{
\donttest{
help(standardize_country_from_iso2)

centroids <- get_centroids(path_centroids=NA)

colnames(centroids)
head(centroids)
}

}
\seealso{
\code{\link[parseGBIF]{prepare_gbif_occurrence_data}}, \code{\link[parseGBIF]{download_gbif_data_from_doi}}
}
\author{
Pablo Hendrigo Alves de Melo,
Nadia Bystriakova &
Alexandre Monro
}
