% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseGBIF_summary.R
\name{parseGBIF_summary}
\alias{parseGBIF_summary}
\title{Selecting the master digital voucher}
\usage{
parseGBIF_summary(
  parseGBIF_all_data = NA,
  file.parseGBIF_all_data = "",
  fields_to_merge = c("Ctrl_fieldNotes", "Ctrl_year", "Ctrl_stateProvince",
    "Ctrl_municipality", "Ctrl_locality", "Ctrl_countryCode", "Ctrl_eventDate",
    "Ctrl_habitat", "Ctrl_level0Name", "Ctrl_level1Name", "Ctrl_level2Name",
    "Ctrl_level3Name")
)
}
\arguments{
\item{parseGBIF_all_data}{...}

\item{file.parseGBIF_all_data}{...}

\item{fields_to_merge}{fields to merge}

\item{fields_to_compare}{fields to compare content frequency}

\item{fields_to_parse}{all fields}

\item{silence}{if TRUE does not display progress messages}
}
\value{
\itemize{
\item \strong{parseGBIF_general_summary}
\item \strong{parseGBIF_merge_fields_summary}
\item \strong{parseGBIF_merge_fields_summary_useable_data}
\item \strong{parseGBIF_merge_fields_summary_unusable_data}
}
}
\description{
...
}
\details{
...
}
\examples{
\donttest{


results <- export_data_v2.1(occ_digital_voucher_file = '',
occ_digital_voucher = occ_digital$all_data,
merge_unusable_data = TRUE,
silence = FALSE)

names(results)

head(results$occ_all)
colnames(results$occ_all)

}
}
\seealso{
\code{\link[ParsGBIF]{batch_checkName_wcvp}}, \code{\link[ParsGBIF]{extract_gbif_issue}}
}
\author{
Pablo Hendrigo Alves de Melo,
Nadia Bystriakova &
Alexandre Monro
}
