% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_gbif_occurrence_data.R
\name{prepare_gbif_occurrence_data}
\alias{prepare_gbif_occurrence_data}
\title{Preparing occurrence data for use by parseGBIF}
\usage{
prepare_gbif_occurrence_data(gbif_occurrece_file = "", columns = "standard")
}
\arguments{
\item{gbif_occurrece_file}{The name of the file from which the with occurrence data
downloaded from GBIF (by default "occurrence.txt")}

\item{columns}{Character vector of strings to indicate column names of the GBIF occurrence file.
Use 'standard' to select basic columns for use in the package, 'all' to select all available columns.
The default is 'standard'}
}
\value{
data.frame with renamed selected fields with prefix "Ctrl_"
}
\description{
Prepare occurrence data downloaded from GBIF to be used by ParsGBIF functions
}
\details{
Select data fields and rename field names prefixed with "Ctrl_"
}
\examples{
\donttest{

library(ParsGBIF)

help(prepare_gbif_occurrence_data)

occ_file <- 'https://raw.githubusercontent.com/pablopains/ParsGBIF/main/dataGBIF/Achatocarpaceae/occurrence.txt'

occ <- prepare_gbif_occurrence_data(gbif_occurrece_file = occ_file,
                                    columns = 'standard')

colnames(occ)

head(occ)
}
}
\seealso{
\code{\link[ParsGBIF]{select_gbif_fields}}, \code{\link[ParsGBIF]{extract_gbif_issue}}
}
\author{
Pablo Hendrigo Alves de Melo,
Nadia Bystriakova &
Alexandre Monro
}
