% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize_country_from_iso2.R
\name{standardize_country_from_iso2}
\alias{standardize_country_from_iso2}
\title{Country names and codes from iso2}
\usage{
standardize_country_from_iso2(
  occ,
  iso2_field_name = "Ctrl_countryCode",
  silence = TRUE
)
}
\arguments{
\item{occ}{GBIF occurrence table with selected columns as select_gbif_fields(columns = 'standard')}

\item{iso2_field_name}{indicates the name of the field with ISO2 code of the countries}
}
\value{
List whith tow data frames, occ, with the original data set plus two columns, parseGBIF_countryCode_ISO3 and
parseGBIF_countryName_en and countrylist, with the list of countries found with all the columns of countrycode::codelist)
}
\description{
Checks and standardizes country names and codes from iso2
}
\details{
Returns the last name
}
\examples{
\donttest{
help(standardize_country_from_iso2)

occ <- prepare_gbif_occurrence_data(gbif_occurrece_file =  'https://raw.githubusercontent.com/pablopains/parseGBIF/main/dataGBIF/Achatocarpaceae/occurrence.txt',
                                    columns = 'standard')
x <- standardize_country_from_iso2(occ = occ,
                                     iso2_field_name = 'Ctrl_countryCode',
                                     return_fields = c('iso3c','country.name.en'))
colnames(x$occ)
head(x$countrycodelist)
}

}
\seealso{
\code{\link[parseGBIF]{prepare_gbif_occurrence_data}}, \code{\link[parseGBIF]{download_gbif_data_from_doi}}
}
\author{
Pablo Hendrigo Alves de Melo,
Nadia Bystriakova &
Alexandre Monro
}
