% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize_scientificName.R
\name{standardize_scientificName}
\alias{standardize_scientificName}
\title{standardize_scientificName}
\usage{
standardize_scientificName(
  searchedName = "Alomia angustata (Gardner) Benth. ex Baker"
)
}
\arguments{
\item{searchedName}{scientific name, with or without author}
}
\value{
searchedName,
standardizeName,
taxonAuthors,
taxonAuthors_last
}
\description{
standardize binomial name, variety, subspecies, form and hybrids, authorship
to allow comparison with names of taxa in the World Checklist of Vascular Plants (WCVP) database
}
\details{
Standardize scientific name according to WCVP format.
Separate generic epithet, specific epithet, variety, subspecies, form, hybrid and author, in the scientific name, if any.
Standardize, according to WCVP, abbreviation of infrataxon, if any:
variety to var.,
subspecies to subsp.,
FORM to f.,
hybrid separator separate x from the specific epithet.
}
\examples{
\donttest{
standardize_scientificName('Leucanthemum ×superbum (Bergmans ex J.W.Ingram) D.H.Kent')
standardize_scientificName('Alomia angustata (Gardner) Benth. ex Baker')
standardize_scientificName('Centaurea ×aemiliae Font Quer')
}

}
\seealso{
\code{\link[ParsGBIF]{get_wcvp}}, \code{\link[ParsGBIF]{checkName_wcvp}}
}
\author{
Pablo Hendrigo Alves de Melo,
Nadia Bystriakova &
Alexandre Monro
}
