% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wcvp_check_name.R
\name{wcvp_check_name}
\alias{wcvp_check_name}
\title{Check species names against World Checklist of Vascular Plants (WCVP) database}
\usage{
wcvp_check_name(
  searchedName = "Hemistylus brasiliensis Wedd.",
  wcvp_names = "",
  if_author_fails_try_without_combinations = TRUE
)
}
\arguments{
\item{searchedName}{scientific name, with or without author}

\item{wcvp_names}{WCVP table, wcvp_names.csv file from http://sftp.kew.org/pub/data-repositories/WCVP/
If NA, automatically load the latest version of the database by the function
parseGBIF::wcvp_get_data(read_only_to_memory = TRUE)$wcvp_names.}

\item{if_author_fails_try_without_combinations}{option for partial verification of the authorship
of the species. Remove the authors of combinations, in parentheses}
}
\value{
Data frame with WCVP fields prefixed with 'wcvp_'
}
\description{
Use the \href{https://powo.science.kew.org/}{World Checklist of Vascular Plants WCVP}
\href{http://sftp.kew.org/pub/data-repositories/WCVP/}{database} to check accepted names and update synonyms.

The World Checklist of Vascular Plants (WCVP) database is available from the
\href{https://powo.science.kew.org/about-wcvp}{Royal Botanic Gardens, Kew}.
It can be downloaded to a folder of the user's choice or into memory using the get_wcvp function. The output has 33 columns.
}
\details{
About the World Checklist of Vascular Plants https://powo.science.kew.org/about-wcvp
searchNotes values:
\itemize{
\item Accepted - When only one authorless scientific name is present in the list of TAXON_name with
and TAXON_STATUS equal to "Accepted", verified_speciesName = 100.
\item Accepted among homonyms - When more than one authorless scientific name is present in the
TAXON_name list, but only one of the homonyms displays TAXON_STATUS equal to "Accepted",
verified_speciesName = number of matches/100.
\item Homonyms - When more than one authorless scientific name is present in the TAXON_name list
and more than one, or none among the homonyms, display TAXON_STATUS equal to "Accepted",
verified_speciesName = number of matches/100.
Before searching for homonyms, there was a failure in trying to find the matching match between
authorless scientific name in TAXON_name and author in TAXON_AUTHORS, in these cases
verified_author equal to 0 (zero),
\item Not Found: When the authorless scientific name is not present in the TAXON_NAME LIST
\item Unplaced: When only one authorless scientific name is present in the list of TAXON_name with and TAXON_STATUS = "Unplaced"
\item Updated: When only one authorless scientific name is present in the list of TAXON_name and ACCEPTED_PLANT_NAME_ID
are not empty (and ACCEPTED_PLANT_NAME_ID is different from the ID of the species consulted)
taxon_status_of_searchedName, plant_name_id_of_searchedName and taxon_authors_of_searchedName values:
\itemize{
\item When searchNotes equals "Updated" - The fields record the information of the scientific name originally consulted.
\item When searchNotes equals "Homonyms" - Fields record the information of homonymous synonyms separated by "|".
}
\item verified_author values:
\itemize{
\item When value equal to 100 - when there is matched match between authorless scientific name in TAXON_name and author in TAXON_AUTHORS.
\item When value equal to 50 - when there is combined correspondence between authorless scientific name in TAXON_name and author, without (combination), in TAXON_AUTHORS.
\item When value equal to 0 - regardless of the correspondence between authorless scientific name in TAXON_name, author is not present in TAXON_AUTHORS.
}
}
}
\examples{
# These examples take >10 seconds to run and require 'parseGBIF::wcvp_get_data()'
\donttest{
library(parseGBIF)

help(wcvp_check_name)

wcvp_names <- wcvp_get_data(read_only_to_memory = TRUE)$wcvp_names

# 1) Updated
wcvp_check_name(searchedName = 'Hemistylus brasiliensis Wedd.',
               wcvp_names = wcvp_names,
               if_author_fails_try_without_combinations = TRUE)

# 2) Accepted
wcvp_check_name(searchedName = 'Hemistylus boehmerioides Wedd. ex Warm.',
               wcvp_names = wcvp_names,
               if_author_fails_try_without_combinations = TRUE)

# 3) Unplaced - taxon_status = Unplaced
wcvp_check_name(searchedName = 'Leucosyke australis Unruh',
               wcvp_names = wcvp_names,
               if_author_fails_try_without_combinations = TRUE)

# 4) Accepted among homonyms - When author is not informed. In this case, one of the homonyms, taxon_status is accepted
wcvp_check_name(searchedName = 'Parietaria cretica',
               wcvp_names = wcvp_names,
               if_author_fails_try_without_combinations = TRUE)

# When author is informed
wcvp_check_name(searchedName = 'Parietaria cretica L.',
               wcvp_names = wcvp_names,
               if_author_fails_try_without_combinations = TRUE)

# When author is informed
wcvp_check_name(searchedName = 'Parietaria cretica Moris',
               wcvp_names = wcvp_names,
               if_author_fails_try_without_combinations = TRUE)

# 5) Homonyms - When author is not informed. In this case, none of the homonyms, taxon_status is Accepted
wcvp_check_name(searchedName = 'Laportea peltata',
               wcvp_names = wcvp_names,
               if_author_fails_try_without_combinations = TRUE)

# When author is informed
wcvp_check_name(searchedName = 'Laportea peltata Gaudich. & Decne.',
               wcvp_names = wcvp_names,
               if_author_fails_try_without_combinations = TRUE)

# When author is informed
wcvp_check_name(searchedName = 'Laportea peltata (Blume) Gaudich.',
               wcvp_names = wcvp_names,
               if_author_fails_try_without_combinations = TRUE)
}
Purpose: Faster wcvp_check_name() with identical output format
Update: Uses wcvp_prepare_index() attributes/keys when present for O(1) name candidate retrieval

}
\seealso{
\code{\link[parseGBIF]{wcvp_check_name_batch}}, \code{\link[parseGBIF]{wcvp_get_data}}
}
\author{
Pablo Hendrigo Alves de Melo,
Nadia Bystriakova &
Alexandre Monro
}
