% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wcvp_check_name_batch.R
\name{wcvp_check_name_batch}
\alias{wcvp_check_name_batch}
\title{In batch, use the WCVP database to check accepted names and update synonyms}
\usage{
wcvp_check_name_batch(
  occ = NA,
  wcvp_names = "",
  if_author_fails_try_without_combinations = TRUE,
  wcvp_selected_fields = "standard",
  silence = TRUE
)
}
\arguments{
\item{occ}{GBIF occurrence table with selected columns as select_gbif_fields(columns = 'standard')}

\item{wcvp_names}{get data frame in parseGBIF::wcvp_get_data(read_only_to_memory = TRUE)$wcvp_names
or configure function to save a copy on local disk to optimize loading, see details in help(wcvp_get_data)}

\item{if_author_fails_try_without_combinations}{option for partial verification of the authorship of the species.
Remove the authors of combinations, in parentheses.}

\item{wcvp_selected_fields}{WCVP fields selected as return, 'standard' basic columns, 'all' all available columns.
The default is 'standard'}

\item{silence}{if TRUE does not display progress messages}
}
\value{
list with two data frames: summary, species list and occ_wcvp_check_name, with WCVP fields
}
\description{
Species’ names can be checked against WCVP database one by one, or in a batch mode.
To verify individual names, the function wcvp_check_name is used.
}
\details{
See help(checkName_wcvp)
\itemize{
\item \href{http://sftp.kew.org/pub/data-repositories/WCVP/}{about WCVP database}
\item \href{https://powo.science.kew.org//}{World Checklist of Vascular Plants}
\item \href{http://sftp.kew.org/pub/data-repositories/WCVP/}{WCVP database}
\item \href{https://powo.science.kew.org/about-wcvp}{(about WCVP)}
}
}
\examples{
# These examples take >10 minutes to run and require 'parseGBIF::wcvp_get_data()'
\donttest{

library(parseGBIF)

help(wcvp_check_name_batch)

occ_file <- 'https://raw.githubusercontent.com/pablopains/parseGBIF/main/dataGBIF/Achatocarpaceae/occurrence.txt'

occ <- prepare_gbif_occurrence_data(gbif_occurrece_file = occ_file,
                                    columns = 'standard')

# wcvp_names <- wcvp_get_data(read_only_to_memory = TRUE)$wcvp_names
data(wcvp_names_Achatocarpaceae)

head(wcvp_names)

res_wcvp_check_name_batch <- wcvp_check_name_batch(occ = occ,
                                                 wcvp_names =  wcvp_names,
                                                 if_author_fails_try_without_combinations = TRUE,
                                                 wcvp_selected_fields = 'standard',
                                                 show_process = TRUE)

names(res_wcvp_check_name_batch)

head(res_wcvp_check_name_batch$summary)

head(res_wcvp_check_name_batch$occ_wcvp_check_name)

}

}
\seealso{
\code{\link[parseGBIF]{wcvp_get_data}}, \code{\link[parseGBIF]{wcvp_check_name}}
}
\author{
Pablo Hendrigo Alves de Melo,
Nadia Bystriakova &
Alexandre Monro
}
